Sample GUI
Here's the sample GUI that I made. Obviously it's a little rough right
now but I'd rather have everyone's approval on a design before we move
forward. Let me know if there's anything you think needs to be added,
removed, or changed and I'll update it as soon as possible.