Adding the first version of the game creator
I ended up using VS 2008 to make this since apparently 2012 doesn't have
forms in c++ anymore and I'm too lazy to mess around with c#. I tried to
keep as much of the clutter out of the repository as I could but if it's
missing something it needs to run, let me know.
As of now, I've tested all the xml tags it generates except for the
conditionals and variables.